BUG:FindITK.cmake removed from CMake (starting in CMake 3.1)
For backward compatibility, FindITK.cmake was setting the variable USE_ITK_FILE to the value of ITK_USE_FILE which is defined in ITKConfig.cmake (in the ITK build directory).
Since FindITK.cmake has been removed from CMake in version 3.1, this variable was not defined anymore and this was creating errors when trying to include the ITK file with 'INCLUDE(${USE_ITK_FILE})'.
The CMakeLists.txt files have been updated to use the new name of the variable (ITK_USE_FILE).