ENH: Warning message is printed when configuring niral_utilities with COMPILE_convertITKformats set to ON
convertITKformats requires the module ITKDeprecated to be built. If the module is not found, CMake prints an error message but that error message is not very specific: It is not obvious for the person configuring the niral_utilities project that the problem comes from convertITKformats. This warning message allows the user to know that convertITKformats is the tool requiring this specific module. The user can then decide to either not compile this module or recompile ITK with this specific module.