ENH: Including SlicerExecutionModel before ITK
When performing:
find_package(SlicerExecutionModel REQUIRED)
include(${SlicerExecutionModel_USE_FILE})
the following lines of code are also called [1]:
find_package(ITK REQUIRED)
include(${ITK_USE_FILE})
This resets the ITK components that are loaded. If one includes ITK with a set of required components beforehand, when including ${ITK_USE_FILE}, ITK generates some C++ files to select the IO components that the factories (image and transform) will be able to handle. If the IO components that are selected with the original call of include(${ITK_USE_FILE}) is different from the IO components that are loaded in include(${SlicerExecutionModel_USE_FILE}), there will be errors when trying to link the project. To avoid any possible problem, we always find ITK after SlicerExecutionModel.

[1] in UseGenerateCLP.cmake